+++ src/footprint/azext_footprint/vendored_sdks/footprint/aio/operations_async/_profile_operations_async.py
@@ -45,9 +45,9 @@ class ProfileOperations:
         self,
         **kwargs
     ) -> AsyncIterable["models.ProfileList"]:
-        """Retrieves the information about all Footprint profiles under a subscription.
+        """Get all Footprint profiles under a subscription.
 
-        Get all Footprint profiles under a subscription.
+        Retrieves the information about all Footprint profiles under a subscription.
 
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either ProfileList or the result of cls(response)
@@ -60,6 +60,10 @@ class ProfileOperations:
         api_version = "2020-02-01-preview"
 
         def prepare_request(next_link=None):
+            # Construct headers
+            header_parameters = {}  # type: Dict[str, Any]
+            header_parameters['Accept'] = 'application/json'
+
             if not next_link:
                 # Construct URL
                 url = self.list_by_subscription.metadata['url']  # type: ignore
@@ -71,15 +75,11 @@ class ProfileOperations:
                 query_parameters = {}  # type: Dict[str, Any]
                 query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
 
+                request = self._client.get(url, query_parameters, header_parameters)
             else:
                 url = next_link
                 query_parameters = {}  # type: Dict[str, Any]
-            # Construct headers
-            header_parameters = {}  # type: Dict[str, Any]
-            header_parameters['Accept'] = 'application/json'
-
-            # Construct and send request
-            request = self._client.get(url, query_parameters, header_parameters)
+                request = self._client.get(url, query_parameters, header_parameters)
             return request
 
         async def extract_data(pipeline_response):
@@ -112,9 +112,9 @@ class ProfileOperations:
         resource_group_name: str,
         **kwargs
     ) -> AsyncIterable["models.ProfileList"]:
-        """Retrieves the information about all Footprint profiles under a resource group.
+        """Get all Footprint profiles under a resource group.
 
-        Get all Footprint profiles under a resource group.
+        Retrieves the information about all Footprint profiles under a resource group.
 
         :param resource_group_name: The name of the resource group.
         :type resource_group_name: str
@@ -129,6 +129,10 @@ class ProfileOperations:
         api_version = "2020-02-01-preview"
 
         def prepare_request(next_link=None):
+            # Construct headers
+            header_parameters = {}  # type: Dict[str, Any]
+            header_parameters['Accept'] = 'application/json'
+
             if not next_link:
                 # Construct URL
                 url = self.list_by_resource_group.metadata['url']  # type: ignore
@@ -141,15 +145,11 @@ class ProfileOperations:
                 query_parameters = {}  # type: Dict[str, Any]
                 query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
 
+                request = self._client.get(url, query_parameters, header_parameters)
             else:
                 url = next_link
                 query_parameters = {}  # type: Dict[str, Any]
-            # Construct headers
-            header_parameters = {}  # type: Dict[str, Any]
-            header_parameters['Accept'] = 'application/json'
-
-            # Construct and send request
-            request = self._client.get(url, query_parameters, header_parameters)
+                request = self._client.get(url, query_parameters, header_parameters)
             return request
 
         async def extract_data(pipeline_response):
@@ -183,9 +183,9 @@ class ProfileOperations:
         profile_name: str,
         **kwargs
     ) -> "models.Profile":
-        """Retrieves the information about a single Footprint profile.
+        """Get a Footprint profile resource.
 
-        Get a Footprint profile resource.
+        Retrieves the information about a single Footprint profile.
 
         :param resource_group_name: The name of the resource group.
         :type resource_group_name: str
@@ -206,7 +206,7 @@ class ProfileOperations:
         path_format_arguments = {
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str', max_length=90, min_length=1),
-            'profileName': self._serialize.url("profile_name", profile_name, 'str', max_length=64, min_length=5, pattern=r'^[a-zA-Z0-9]'),
+            'profileName': self._serialize.url("profile_name", profile_name, 'str', max_length=64, min_length=3, pattern=r'^[a-zA-Z0-9]'),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -218,7 +218,6 @@ class ProfileOperations:
         header_parameters = {}  # type: Dict[str, Any]
         header_parameters['Accept'] = 'application/json'
 
-        # Construct and send request
         request = self._client.get(url, query_parameters, header_parameters)
         pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
         response = pipeline_response.http_response
@@ -249,9 +248,9 @@ class ProfileOperations:
         reporting_endpoints: Optional[List[str]] = None,
         **kwargs
     ) -> "models.Profile":
-        """Creates or updates a Footprint profile with the specified properties.
+        """Creates or updates a Footprint profile resource.
 
-        Creates or updates a Footprint profile resource.
+        Creates or updates a Footprint profile with the specified properties.
 
         :param resource_group_name: The name of the resource group.
         :type resource_group_name: str
@@ -291,7 +290,7 @@ class ProfileOperations:
         path_format_arguments = {
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str', max_length=90, min_length=1),
-            'profileName': self._serialize.url("profile_name", profile_name, 'str', max_length=64, min_length=5, pattern=r'^[a-zA-Z0-9]'),
+            'profileName': self._serialize.url("profile_name", profile_name, 'str', max_length=64, min_length=3, pattern=r'^[a-zA-Z0-9]'),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -304,7 +303,6 @@ class ProfileOperations:
         header_parameters['Content-Type'] = self._serialize.header("content_type", content_type, 'str')
         header_parameters['Accept'] = 'application/json'
 
-        # Construct and send request
         body_content_kwargs = {}  # type: Dict[str, Any]
         body_content = self._serialize.body(_parameters, 'Profile')
         body_content_kwargs['content'] = body_content
@@ -318,7 +316,6 @@ class ProfileOperations:
             error = self._deserialize(models.DefaultErrorResponse, response)
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
-        deserialized = None
         if response.status_code == 200:
             deserialized = self._deserialize('Profile', pipeline_response)
 
@@ -337,9 +334,9 @@ class ProfileOperations:
         profile_name: str,
         **kwargs
     ) -> None:
-        """Deletes an existing Footprint profile.
+        """Deletes a Footprint profile resource.
 
-        Deletes a Footprint profile resource.
+        Deletes an existing Footprint profile.
 
         :param resource_group_name: The name of the resource group.
         :type resource_group_name: str
@@ -360,7 +357,7 @@ class ProfileOperations:
         path_format_arguments = {
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str', max_length=90, min_length=1),
-            'profileName': self._serialize.url("profile_name", profile_name, 'str', max_length=64, min_length=5, pattern=r'^[a-zA-Z0-9]'),
+            'profileName': self._serialize.url("profile_name", profile_name, 'str', max_length=64, min_length=3, pattern=r'^[a-zA-Z0-9]'),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -371,7 +368,6 @@ class ProfileOperations:
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
 
-        # Construct and send request
         request = self._client.delete(url, query_parameters, header_parameters)
         pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
         response = pipeline_response.http_response
@@ -393,9 +389,9 @@ class ProfileOperations:
         tags: Optional[Dict[str, str]] = None,
         **kwargs
     ) -> "models.Profile":
-        """Updates an existing Footprint profile resource.
+        """Updates a Footprint profile resource.
 
-        Updates a Footprint profile resource.
+        Updates an existing Footprint profile resource.
 
         :param resource_group_name: The name of the resource group.
         :type resource_group_name: str
@@ -421,7 +417,7 @@ class ProfileOperations:
         path_format_arguments = {
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str', max_length=90, min_length=1),
-            'profileName': self._serialize.url("profile_name", profile_name, 'str', max_length=64, min_length=5, pattern=r'^[a-zA-Z0-9]'),
+            'profileName': self._serialize.url("profile_name", profile_name, 'str', max_length=64, min_length=3, pattern=r'^[a-zA-Z0-9]'),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -434,7 +430,6 @@ class ProfileOperations:
         header_parameters['Content-Type'] = self._serialize.header("content_type", content_type, 'str')
         header_parameters['Accept'] = 'application/json'
 
-        # Construct and send request
         body_content_kwargs = {}  # type: Dict[str, Any]
         body_content = self._serialize.body(_parameters, 'ProfilePatch')
         body_content_kwargs['content'] = body_content