+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,19 @@
 
 Release History
 ===============
+0.4.63
++++++
+* Enable the September (2020-09-01) for use with the AKS commands
+* Support Start/Stop cluster feature in preview
+* Support ephemeral OS functionality
+* Add new properties to the autoscaler profile: max-empty-bulk-delete, skip-nodes-with-local-storage, skip-nodes-with-system-pods, expander, max-total-unready-percentage, ok-total-unready-count and new-pod-scale-up-delay
+* Fix case sensitive issue for AKS dashboard addon
+* Remove PREVIEW from azure policy addon
+
+0.4.62
++++++
+* Add support for enable/disable confcom (sgx) addon.
+
 0.4.61
 +++++
 * Fix AGIC typo and remove preview label from VN #2141